move the fake server
